+++ src/main/scala/com/fotolog/redis/RedisClientTypes.scala
@@ -1,0 +1,13 @@
+package com.fotolog.redis
+
+/**
+ * Created with IntelliJ IDEA.
+ * User: sergey
+ * Date: 10/29/13
+ * Time: 9:30 AM
+ * To change this template use File | Settings | File Templates.
+ */
+object RedisClientTypes {
+  type BinVal = Array[Byte]
+  type KV = Tuple2[String, BinVal]
+}
